\name{bgw-internal}
\alias{divset_c}
\alias{drglg_c}
\value{
These functions are wrappers for 'Fortran' routines, and are not meant to be called by the user.
}
\title{Internal BGW Functions}
\description{
 Internal BGW functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
