% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgw_checkSetting.R
\name{bgw_checkSetting}
\alias{bgw_checkSetting}
\title{bgw_checkSetting}
\usage{
bgw_checkSetting(
  settingValue,
  bgw_setting_type,
  bgw_validDiscrete,
  bgw_contLB,
  bgw_contUB
)
}
\arguments{
\item{settingValue}{Setting value (submitted by user).}

\item{bgw_setting_type}{Type of setting being checked. Possible values are "discrete" and "continuous".}

\item{bgw_validDiscrete}{List. Contains valid values for a discrete setting.}

\item{bgw_contLB}{Numerical value. The lower bound for a valid continuous setting.}

\item{bgw_contUB}{Numerical value. The upper bound for a valid continuous setting.}
}
\value{
Logical. Indicates if the setting is okay or not.
}
\description{
Checks to see if user-provided value for a bgw_setting is valid.
}
