% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgw_drglg.R
\name{bgw_drglg}
\alias{bgw_drglg}
\title{bgw_drglg}
\usage{
bgw_drglg(
  d,
  dr,
  iv,
  liv,
  lv,
  n,
  nd,
  nn,
  p,
  ps,
  r,
  rd,
  v,
  x,
  rhoi,
  rhor,
  i_itsum
)
}
\arguments{
\item{d}{Scaling vector}

\item{dr}{Derivative of the choice probability model wrt x}

\item{iv}{BGW internal vector of integer values}

\item{liv}{Length of iv.}

\item{lv}{Length of v.}

\item{n}{Dimension of vector (r) of generalized residuals for the model}

\item{nd}{Leading dimension dr. Must be at least ps.}

\item{nn}{Leading dimension of r, rd}

\item{p}{Dimension of x (as well as d, g) = number of parameters being estimated}

\item{ps}{Number of non-nuisance parameters (= p in this implementation)}

\item{r}{Vector of generalized residuals for the model}

\item{rd}{Vector of storage space for regression diagonostics (not currently used)}

\item{v}{BGW internal vector of numeric values}

\item{x}{Parameter vector for which the objective function is being minimized}

\item{rhoi}{Vector of integers for use by user (not currently used)}

\item{rhor}{Vector of numeric values for use by user (not currently used)}

\item{i_itsum}{Variable for passing itsum instruction back to bgw_mle}
}
\value{
out List of return values.
}
\description{
An r wrapper to call drglg_c, which in turn is a wrapper
  to call the Fortran subroutine drglg.
  drglg is the BGW iteration driver for performing statistical parameter estimation
}
