% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgw_itsum.R
\name{bgw_itsum}
\alias{bgw_itsum}
\title{bgw_itsum}
\usage{
bgw_itsum(d, g, iv, v, x, p, betaIsNamed, betaNames = NULL, i_itsum)
}
\arguments{
\item{d}{Scaling vector}

\item{g}{Gradient of objective function (negative-log-likelihood) wrt x}

\item{iv}{BGW internal vector of integer values}

\item{v}{BGW internal vector of numeric values}

\item{x}{Parameter vector for which the objective function is being minimized}

\item{p}{Dimension of x}

\item{betaIsNamed}{Logical.}

\item{betaNames}{Character vector. If available, has beta parameter names.}

\item{i_itsum}{Code from caller to select specific options}
}
\value{
iv There are iv values changed inside bgw_itsum. The iv vector is the integer workspace for Fortran BGW.
}
\description{
Prints iteration summary, info on initial and final x.
}
