% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgw_writeIterations.R
\name{bgw_writeIterations}
\alias{bgw_writeIterations}
\title{Writes the vector [beta,ll] to a file called \code{modelname_iterations.csv}
#
Was created using apollo_writeTheta as a starting point...
Because this is an internal function, the inputs will be assumed to be clean.}
\usage{
bgw_writeIterations(beta, ll, outputFile)
}
\arguments{
\item{beta}{vector of parameters to be written (for now, no fixed betas).}

\item{ll}{scalar representing the log-likelihood of the whole model.}

\item{outputFile}{Character. Name of the output file.}
}
\value{
Nothing.
}
\description{
Writes the vector [beta,ll] to a file called \code{modelname_iterations.csv}
#
Was created using apollo_writeTheta as a starting point...
Because this is an internal function, the inputs will be assumed to be clean.
}
